Move computeNumberOfStars into dedicated rating-utils file
https://coveord.atlassian.net/browse/KIT-5000
